+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added support for Delphi 13 `noreturn` directive.
 - Added support for Delphi 13 `unmanaged` generic constraint.
+- Added support for `if else` ternary expressions.
 
 ## [0.7.0] - 2025-11-11